requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==1.11
django-bootstrap4==0.0.7
django-chartjs==1.3
idna==2.7
nltk==3.3
oauthlib==2.1.0
PySocks==1.6.8
pytz==2018.7
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
textblob==0.15.1
tweepy==3.6.0
urllib3==1.24
